feat(bal-time-input): extracted formatTime function to util
